Add login, logout, current

+++ src/services/contacts.service.ts
@@ -1,14 +1,14 @@
 import { ContactModel } from 'models/contact.schema';
 import { ContactType } from 'types/Contact.type';
 
-export const getContactsService = async () => {
-  const contacts = await ContactModel.find();
+export const getContactsService = async (owner: string) => {
+  const contacts = await ContactModel.find({ owner }).select({ __v: 0 });
 
   return contacts;
 };
 
-export const getContactByIdService = async (contactId: string) => {
-  const contacts = await ContactModel.findById(contactId);
+export const getContactByIdService = async (contactId: string, owner: string) => {
+  const contacts = await ContactModel.findOne({ _id: contactId, owner }).select({ __v: 0 });
 
   return contacts;
 };
@@ -19,14 +19,16 @@ export const addContactService = async (contact: ContactType) => {
   return newContact;
 };
 
-export const removeContactByIdService = async (contactId: string) => {
-  const removedContact = await ContactModel.findByIdAndRemove({ _id: contactId });
+export const removeContactByIdService = async (contactId: string, owner: string) => {
+  const removedContact = await ContactModel.findByIdAndRemove({ _id: contactId, owner }).select({ __v: 0 });
 
   return removedContact;
 };
 
-export const updateContactByIdService = async (contactId: string, body: Partial<ContactType>) => {
-  const updatedContact = await ContactModel.findByIdAndUpdate({ _id: contactId }, body, { new: true });
+export const updateContactByIdService = async (contactId: string, body: Partial<ContactType>, owner: string) => {
+  const updatedContact = await ContactModel.findByIdAndUpdate({ _id: contactId, owner }, body, { new: true }).select({
+    __v: 0,
+  });
 
   return updatedContact;
 };
